Auto-route: an option to follow the channel markers (#56)
The depth model routes on the water — it knows where it is deep, and among
safe routes it prefers the deeper one. It does not know where the channel
is. The marks do: they carry the dredging, the tide, the shifting bar and
the traffic that a DEPARE polygon compiled years ago does not.

`?channel_markers=1` on /autoroute (and `follow_channel_markers` on
/optimize) reconstructs the buoyed channel from the charted marks and
prefers it. A port-hand and a starboard-hand mark that are each other's
nearest opposite-handed neighbour, within 800 m and with no charted land
between, are a GATE: the water you pass between. A safe-water mark is a
mid-channel gate on its own. Gate midpoints are linked to their nearest
neighbours to recover the centreline, and being beside that channel rather
than in it costs more.

It is expressed as a cost outside the channel, not a discount inside it.
The A* heuristic is straight-line metres at multiplier 1, so a multiplier
below 1 would make it inadmissible; and a discount would pull a route
toward any buoy anywhere, including ones marking something that is not a
channel. The cost is local — it prices only the water within ~250 m of a
channel the marks actually gate — so an offshore passage is untouched, and
it can never open water the depth model closed or close water it left open.

A mark with no partner is ignored. On its own a lateral mark says which
side to pass it, and without a heading that is not a channel; penalising
the water around a lone buoy would push a route away from it in every
direction, including down the channel it marks. When nothing was charted to
follow, the result says so in `warnings` rather than letting the operator
believe the buoys shaped the route.

The marks ride along on the all-band channel query, because a buoy is
charted in the harbour cells the resolution ceiling drops. A grid built
from precomputed nav tiles gets them from a small dedicated query instead:
the tile flag byte is a stored format that is already full, and baking a
per-request preference into every tile on disk is the wrong trade.

Cell flags widen to uint16 to make room for the two new ones. The stored
NavTile format is unchanged — storedFlags/liveFlags still map to and from
the same uint8.

Clients: a "Follow the channel markers" checkbox in the web Routes panel,
wired into both Plan and Optimize. The mobile client carries the parameter
so the two stay in sync on the contract; the touch design for a toggle in
the auto-route sheet is raised as card E6.

Go build and test could not run here — this session cannot fetch the Go
1.25 toolchain (storage.googleapis.com is denied by the egress policy), so
the Go changes are gofmt-clean and reviewed but unexecuted. Web tests and
lint pass.

+++ render/navgrid_test.go
@@ -110,7 +110,7 @@ func TestSampleTilesReportsNoCoverage(t *testing.T) {
 }
 
 func TestStoredAndLiveFlagsRoundTrip(t *testing.T) {
-	for _, live := range []uint8{
+	for _, live := range []uint16{
 		cellLand, cellObstruction, cellDredged, cellUnsurveyed, cellRestricted,
 		cellLand | cellRestricted, cellDredged | cellUnsurveyed,
 	} {
